ad9656: add reference project for the ad9656 eval board
Added new example application for ad9656 evaluation board coupled with the
zcu102 carrier. The application uses the evaluation board's clock of 125MHz
to clock the design. The ad9508 and ad9553 clock chips are configured such
that the clock recevied as the reference clock for ad9656 chip has 125MHz.
The JESD204 interface that transfers data from the ad9656 chip to the FPGA
is configured at a line rate of 2.5GHz. The application first tests if the
JESD204 connection is working properly and then it programs the ad9656 chip
with a user test pattern to be sent to the FPGA instead of the real ADC
data.

Signed-off-by: Dan Hotoleanu <dan.hotoleanu@analog.com>